requirements.txt
configobj==5.0.9
gitdb==4.0.12
GitPython==3.1.44
joblib==1.4.2
smmap==5.0.2
streams.py==1.3.0
